# swagger.yaml
openapi: 3.0.0
info:
  title: Freebird CV Theque API
  description: Tableau de bord recrutement, un outil de pilotage en ressources humaines
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /offres/{id}:
    parameters:
      - in: path
        name: id
        description: offre id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single offre by id
      description: Admin users can access any offre, regular users are restricted to their own offre.
      operationId: getoffreById
      security:
        - bearerAuth: []
      
    put:
      summary: Update an offre
      description: Admin users can update any offre including role, regular users are restricted to their own offre and cannot update role.
      operationId: updateoffre
      security:
        - bearerAuth: []     
      
    delete:
      summary: Delete an offre
      description: Admin users can delete any offre, regular users are restricted to their own offre.
      operationId: deleteoffre
      security:
        - bearerAuth: []    
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...